requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
cryptography==41.0.4
defusedxml==0.8.0rc2
Django==4.2.5
django-filter==23.3
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
drf-extra-fields==3.7.0
filetype==1.2.0
idna==3.4
itypes==1.2.0
mccabe==0.6.1
oauthlib==3.2.2
Pillow==10.0.0
psycopg2-binary==2.9.8
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.1.0
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2021.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sqlparse==0.3.1
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.5
